requirements.txt
requests>=2.28.0
watchdog>=3.0.0
python-dotenv>=1.0.0
opencc-python-reimplemented>=0.1.7
